+++ backend/models/userModel.js
@@ -1,24 +1,26 @@
-import mongoose from 'mongoose';
+const mongoose = require('mongoose');
+const crypto = require('crypto'); // Import the crypto module
 
 const userSchema = new mongoose.Schema({
     name: {
         type: String,
-        required: true
+        required: [true, 'Please add a name']
     },
     email: {
         type: String,
-        required: true,
-        unique: true
+        unique: true,
+        required: [true, 'Please add an email']
     },
     password: {
         type: String,
-        required: true
+        required: [true, 'Please add a password']
     },
     accessToken: {
-        type: String
+        type: String,
+        default: () => crypto.randomBytes(128).toString('hex')
     }
-});
+},
+    { timestamps: true }); // Fix the typo in timestamps
 
 const User = mongoose.model('User', userSchema);
-
-export default User;
+module.exports = User;
