--- backend/routes/sessionRoutes.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import { loginUser } from '../controllers/sessionController';
-
-const router = express.Router();
-
-router.post('/sessions', loginUser);
-
-export default router;
